fix: fix a regression bug of rendering the toast icon by appearance
